# perf: url encode path segments in longer string slices