--- a/url/benches/parse_url.rs
+++ b/url/benches/parse_url.rs
@@ -13,7 +13,7 @@ fn short(bench: &mut Bencher) {
 }
 
 fn long(bench: &mut Bencher) {
-    let url = "https://example.com/parkbench?tre=es&st=uff";
+    let url = "https://example.com/parkbench?tre=es&st=ufflongerlonger";
 
     bench.bytes = url.len() as u64;
     bench.iter(|| black_box(url).parse::<Url>().unwrap());