Before:

```
test long           ... bench:         250 ns/iter (+/- 11) = 172 MB/s
```

After:

```
test long           ... bench:         221 ns/iter (+/- 24) = 248 MB/s
```