@Scripter17 2025-02-08
It fully occurred to me how annoying it must be to get a "you missed a smudge" review from a rando on your genuinely very impressive work so I want to redeem myself: You're awesome

URL parsing has been a bit of a bottleneck in my current main project for a while and all your work does a pretty good dent in it

Seeing all your pull requests was a very nice thing to wake up and look forward to